Subset handling and exported functions
- Added subset parameter to InitializeWordBag so that filtering can be
accomodated natively in this function. Previously it had incorrectly
relied on the user filtering the text before supplying it. Filtered
responses do not impact the processing done by the word bag, however the
wordbag does apply the processing to all responses.
- Information about the filter and about how many cases are blank is now
displayed in the DT widget table by print.wordBag.
- Exported functions for saving sentiment and tidied text so that the R
code used in Q/Displayr is more transparent. The new functions are
SaveTidiedText and SaveNetSentimentScores.
- Added unit tests for the above changes.
- Improved documentation.